refactor(ledger): move ConwayProtocolParams from test to main
Lifts the case class, its CBOR `Decoder`, `fromCbor`, and `toProtocolParams`
out of `scalus.testing.conformance.LedgerState` (test source set) into a new
`scalus.cardano.ledger.ConwayProtocolParams` in main. This makes the Conway
PParams CBOR shape — what `cardano-ledger`'s `EncCBOR (PParams Conway)`
produces and the on-disk `pparams-by-hash` files contain — usable by
production callers.

Motivating consumer: a `LocalStateQuery` driver in scalus-node decoding the
result of the `GetCurrentPParams` query needs `Decoder[ConwayProtocolParams]`
plus `.toProtocolParams` to surface a fully populated `ProtocolParams` to
`BackupSource.LocalNode.fetchLatestParams`. Today that decoder lives in test
scope, so the production path can't reach it.

The test-fixture helpers `loadFromHash` / `extractPparamsHash` (which index
the conformance test vectors' on-disk pparams-by-hash directory) stay in
test code under a new `ConwayProtocolParamsFixtures` object — they're
fixture-loading utilities, not protocol code. Existing callers updated.

No behavioural change. All 763 scalus-cardano-ledger tests pass; the
conformance integration test suite (scalusCardanoLedgerIt) compiles unchanged.
